+++ src/compas_model/elements/__init__.py
@@ -4,8 +4,6 @@ from .element import reset_computed
 from .block import BlockElement
 from .block import BlockFeature
 from .block import BlockGeometry
-from .interface import InterfaceElement
-from .interface import InterfaceFeature
 from .plate import PlateElement
 from .plate import PlateFeature
 
@@ -17,8 +15,6 @@ __all__ = [
     "BlockElement",
     "BlockFeature",
     "BlockGeometry",
-    "InterfaceElement",
-    "InterfaceFeature",
     "PlateElement",
     "PlateFeature",
 ]