+++ src/compas_model/materials/steel.py
@@ -1,8 +1,11 @@
+from typing import Literal
+from typing import Optional
+
 from .material import Material
 
 
 class Steel(Material):
-    strength_classes = {
+    STRENGTH_CLASSES: dict[str, dict[str, float]] = {
         "S235": {"fy": 235, "fu": 360},
         "S275": {"fy": 275, "fu": 430},
         "S355": {"fy": 355, "fu": 490},
@@ -10,9 +13,8 @@ class Steel(Material):
     }
 
     @property
-    def __data__(self):
-        # type: () -> dict
-        data = super(Steel, self).__data__
+    def __data__(self) -> dict:
+        data = super().__data__
         data.update(
             {
                 "fy": self.fy,
@@ -21,8 +23,9 @@ class Steel(Material):
         )
         return data
 
-    def __init__(self, fy, fu, name=None):
-        super(Steel, self).__init__(name=name)
+    def __init__(self, fy: float, fu: float, name: Optional[str] = None) -> None:
+        super().__init__(name=name)
+
         self.fy = fy
         self.fu = fu
         self.E = 210
@@ -30,20 +33,19 @@ class Steel(Material):
         self.density = 7850
 
     @property
-    def rho(self):
+    def rho(self) -> float:
         return self.density
 
     @property
-    def nu(self):
+    def nu(self) -> float:
         return self.poisson
 
     @property
-    def G(self):
+    def G(self) -> float:
         return self.E / (2 * (1 + self.nu))
 
     @classmethod
-    def from_strength_class(cls, strength_class):
-        # type: (str) -> Steel
+    def from_strength_class(cls, strength_class: Literal["S235", "S275", "S355", "S450"]) -> "Steel":
         """Construct a steel material from a steel strength class.
 
         Parameters
@@ -57,7 +59,7 @@ class Steel(Material):
 
         """
         strength_class = strength_class.upper()
-        if strength_class not in cls.strength_classes:
+        if strength_class not in cls.STRENGTH_CLASSES:
             raise ValueError("This strength class is not supported: {}".format(strength_class))
-        params = cls.strength_classes[strength_class]
+        params = cls.STRENGTH_CLASSES[strength_class]
         return cls(**params)